Version 1.04a - Added ThemeInfo (with author and description) and now stored in new folder hierarchie. - Fixed: Typing while UI is Open could switch accidently theme. - Added OverlayMap functions for Special Keys like NumLockActive or CapsLockActive - XmlConfig now better reports errors and refactored map parsing - SwitchMap detection is now faster    - BoardAnimations does not crash anymore while user spamming keys.  - Animation Ticks taking care of still running while triggered

+++ source/K70Tray/BoardAnimation.cpp
@@ -68,7 +68,7 @@ void BoardAnimation::startAt(int x, int y) {
 
 void BoardAnimation::tick() {
 	// DebugMsg("BoardAnimation Tick at %i with size of %i", currentTick, placedAnimation.size());
-	int animationEndCount = 0;
+	size_t animationEndCount = 0;
 	for (map <string, vector<K70RGB>>::iterator it = placedAnimation.begin(); it != placedAnimation.end(); ++it) {		
 		int x = getPositionOfAnimationAdress(it->first, "x") * 4;
 		int y = getPositionOfAnimationAdress(it->first, "y");
